feat: Add ClientRootLayout to Page component
